requirements.txt
contourpy==1.2.0
cycler==0.12.1
fonttools==4.50.0
imageio==2.34.0
kiwisolver==1.4.5
lazy_loader==0.3
matplotlib==3.8.3
networkx==3.2.1
numpy==1.26.4
packaging==24.0
pandas==2.2.1
pillow==10.2.0
pyparsing==3.1.2
pytesseract==0.3.10
python-dateutil==2.9.0.post0
pytz==2024.1
scikit-image==0.22.0
scipy==1.12.0
six==1.16.0
tifffile==2024.2.12
tzdata==2024.1
